+++ src/Support/BlogDatabaseConnection.php
@@ -1,0 +1,21 @@
+<?php
+
+declare(strict_types=1);
+
+namespace CmsOrbit\Blog\Support;
+
+use CmsOrbit\Saas\Support\HostConnection;
+
+class BlogDatabaseConnection
+{
+    public static function name(): string
+    {
+        $connection = config('blog.database.connection');
+
+        if (is_string($connection) && $connection !== '') {
+            return $connection;
+        }
+
+        return HostConnection::name();
+    }
+}